Move the Education.jsx to a separate 'education' folder.
